GUI first implementation done

+++ GUI/Components/Entires/SpecimenCountGUI.py
@@ -1,10 +1,10 @@
-from tkinter import IntVar, Frame
+from tkinter import DoubleVar, Frame
 
 from GUI.Components.Base.EntryGUI import EntryGUI
 
 
 class SpecimenCountGUI(EntryGUI):
-    def __init__(self, state: IntVar, root: Frame, label: str):
+    def __init__(self, state: DoubleVar, root: Frame, label: str):
         super().__init__(state, root, label)
 
     def _renderCheckbox(self):
